requirements.txt
sentence-transformers==4.0.2
faiss-cpu==1.10.0
numpy==2.2.4
PyPDF2==3.0.1
tqdm==4.67.1
nltk==3.8.1
scikit-learn==1.6.1
scipy==1.15.2
torch==2.6.0
transformers==4.50.3
huggingface-hub==0.30.1
tokenizers==0.21.1
safetensors==0.5.3